feat: 添加macos支持 (#1)
* 1

* 2

* 3

* 4

* 5

* 6

* 7

* 8

* 9

* 10

* 11

* 12

* 13

* 14

* 15

* 16

* 17

* 18

* 19

* 20

* 21

* 22

* 23

* 24

* 25

* 26

* 27

* 28

* 29

* 30

* 31

* 32

* 33

* 34

* 35

* 36

* 37

* 38

* 39

* 40

* 41

* 42

* 43

* 44

* 45

* 46

* 47

* 48

* 49

* 50

* 51

* 52

* 53

* 54

* 55

* 56

* 57

* 58

+++ mbnet/websocket/SocketStreamHandle.h
@@ -120,7 +120,7 @@ private:
     };
 
     uintptr_t m_workerThread;
-    long m_stopThread;
+    int m_stopThread;
     WTF::RecursiveMutex m_mutexSend;
     WTF::RecursiveMutex m_mutexReceive;
     WTF::Deque<SocketData*> m_sendData;
